[subgraph] Remove support for standalone articles
All articles must be posted in a publication
