+++ packages/subgraph/schema.graphql
@@ -1,6 +1,6 @@
 type Article @entity {
   id: ID!
-  publication: Publication
+  publication: Publication!
   poster: Bytes! # message sender for article creation event
   article: String! # markdown formatted string or IPFS hash
   authors: [String!] # array of authors (address or string name)
