+++ packages/subgraph/tests/article.test.ts
@@ -10,111 +10,179 @@ import {
 } from "matchstick-as/assembly/index"
 import { getArticleId } from "../src/article.mapping"
 import { handleNewPost } from "../src/mapping"
-import { ARTICLE_ENTITY_TYPE, createNewPostEvent, PUBLICATION_TAG } from "./util"
+import { getPublicationId } from "../src/publication.mapping"
+import { ARTICLE_ENTITY_TYPE, createNewPostEvent, PUBLICATION_ENTITY_TYPE, PUBLICATION_TAG } from "./util"
 
-test("An account can create a standalone article", () => {
+test("An account can NOTE create an article without specifying the publication.", () => {
   const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
-  const action = "article/create"
-  const title = "My First Blog Post"
+  const articleTitle = "My First Blog Post"
   const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
-  const content = `{
-    "action": "${action}",
-    "article": "${article}",
-    "title": "${title}"
-  }`
-
-  const newPostEvent = createNewPostEvent(user, content, PUBLICATION_TAG)
-  const articleId = getArticleId(newPostEvent)
-  handleNewPost(newPostEvent)
-
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
-  assert.notInStore(ARTICLE_ENTITY_TYPE, articleId + "_fake")
-  clearStore()
-})
-
-test("An account can delete a standalone article that it has created", () => {
-  const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
-  const title = "My First Blog Post"
-  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
-  const createArticleContent = `{
+  const articleContent = `{
     "action": "article/create",
     "article": "${article}",
-    "title": "${title}"
+    "title": "${articleTitle}"
   }`
 
-  const createArticlePostEvent = createNewPostEvent(user, createArticleContent, PUBLICATION_TAG)
-  const articleId = getArticleId(createArticlePostEvent)
-  handleNewPost(createArticlePostEvent)
+  const newArticlePostEvent = createNewPostEvent(user, articleContent, PUBLICATION_TAG)
+  const articleId = getArticleId(newArticlePostEvent)
+  handleNewPost(newArticlePostEvent)
 
-  // check that the article is in the store
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
-
-  const deleteArticleContent = `{
-    "action": "article/delete",
-    "id": "${articleId}"
-  }`
-
-  const newPostEvent = createNewPostEvent(user, deleteArticleContent, PUBLICATION_TAG)
-  handleNewPost(newPostEvent)
-
-  // check the article is deleted from the store
   assert.notInStore(ARTICLE_ENTITY_TYPE, articleId)
+
   clearStore()
 })
 
-test("An account can NOT delete a standalone article that is created by another account", () => {
-  const messageSender = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
-  const title = "My First Blog Post"
-  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
-  const createArticleContent = `{
-    "action": "article/create",
-    "article": "${article}",
-    "title": "${title}"
+test("An account can create a article in a publication (where the account has `article/create` permissions)", () => {
+  const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
+  const publicationTitle = "My First Publication"
+  const publicationContent = `{
+    "action": "publication/create",
+    "title": "${publicationTitle}"
   }`
 
-  const createArticlePostEvent = createNewPostEvent(messageSender, createArticleContent, PUBLICATION_TAG)
-  const articleId = getArticleId(createArticlePostEvent)
-  handleNewPost(createArticlePostEvent)
+  const newPublicationPostEvent = createNewPostEvent(user, publicationContent, PUBLICATION_TAG)
+  const publicationId = getPublicationId(newPublicationPostEvent)
+  handleNewPost(newPublicationPostEvent)
 
-  // check that the article is in the store
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
+  assert.fieldEquals(PUBLICATION_ENTITY_TYPE, publicationId, "title", publicationTitle)
+  assert.fieldEquals(PUBLICATION_ENTITY_TYPE, publicationId, "id", publicationId)
+
+  const articleTitle = "My First Blog Post"
+  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
+  const articleContent = `{
+    "action": "article/create",
+    "publicationId": "${publicationId}",
+    "article": "${article}",
+    "title": "${articleTitle}"
+  }`
+
+  const newArticlePostEvent = createNewPostEvent(user, articleContent, PUBLICATION_TAG)
+  const articleId = getArticleId(newArticlePostEvent)
+  handleNewPost(newArticlePostEvent)
+
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
   assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "publication", publicationId)
 
-  const deleteArticleContent = `{
+  assert.notInStore(ARTICLE_ENTITY_TYPE, articleId + "_fake")
+
+  clearStore()
+})
+
+test("An account can delete an article in a publication (where the account has `article/delete` permissions)", () => {
+  const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
+  const publicationTitle = "My First Publication"
+  const publicationContent = `{
+    "action": "publication/create",
+    "title": "${publicationTitle}"
+  }`
+
+  const newPublicationPostEvent = createNewPostEvent(user, publicationContent, PUBLICATION_TAG)
+  const publicationId = getPublicationId(newPublicationPostEvent)
+  handleNewPost(newPublicationPostEvent)
+
+  assert.fieldEquals(PUBLICATION_ENTITY_TYPE, publicationId, "id", publicationId)
+
+  const articleTitle = "My First Blog Post"
+  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
+  const articleContent = `{
+    "action": "article/create",
+    "publicationId": "${publicationId}",
+    "article": "${article}",
+    "title": "${articleTitle}"
+  }`
+
+  const newArticlePostEvent = createNewPostEvent(user, articleContent, PUBLICATION_TAG)
+  const articleId = getArticleId(newArticlePostEvent)
+  handleNewPost(newArticlePostEvent)
+
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
+
+  const articleDeleteContent = `{
     "action": "article/delete",
     "id": "${articleId}"
   }`
 
-  const anotherMessageSender = Address.fromString("0xD028d504316FEc029CFa36bdc3A8f053F6E5a6e4")
+  const newArticleDeletePostEvent = createNewPostEvent(user, articleDeleteContent, PUBLICATION_TAG)
+  handleNewPost(newArticleDeletePostEvent)
 
-  const newPostEvent = createNewPostEvent(anotherMessageSender, deleteArticleContent, PUBLICATION_TAG)
-  handleNewPost(newPostEvent)
-
-  // check the article is NOT deleted
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
+  assert.notInStore(ARTICLE_ENTITY_TYPE, articleId)
 
   clearStore()
 })
 
-test("An account can update a standalone article that it has created", () => {
+test("An account can NOT an delete an article in a publication (where the account do NOT have `article/delete` permissions)", () => {
   const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
-  const title = "My First Blog Post"
-  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
-  const createArticleContent = `{
-    "action": "article/create",
-    "article": "${article}",
-    "title": "${title}"
+  const publicationTitle = "My First Publication"
+  const publicationContent = `{
+    "action": "publication/create",
+    "title": "${publicationTitle}"
   }`
 
-  const createArticlePostEvent = createNewPostEvent(user, createArticleContent, PUBLICATION_TAG)
-  const articleId = getArticleId(createArticlePostEvent)
-  handleNewPost(createArticlePostEvent)
+  const newPublicationPostEvent = createNewPostEvent(user, publicationContent, PUBLICATION_TAG)
+  const publicationId = getPublicationId(newPublicationPostEvent)
+  handleNewPost(newPublicationPostEvent)
 
-  // check that the article is in the store
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
+  assert.fieldEquals(PUBLICATION_ENTITY_TYPE, publicationId, "id", publicationId)
+
+  const articleTitle = "My First Blog Post"
+  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
+  const articleContent = `{
+    "action": "article/create",
+    "publicationId": "${publicationId}",
+    "article": "${article}",
+    "title": "${articleTitle}"
+  }`
+
+  const newArticlePostEvent = createNewPostEvent(user, articleContent, PUBLICATION_TAG)
+  const articleId = getArticleId(newArticlePostEvent)
+  handleNewPost(newArticlePostEvent)
+
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
+
+  const otherUser = Address.fromString("0xD028d504316FEc029CFa36bdc3A8f053F6E5a6e4")
+  const articleDeleteContent = `{
+    "action": "article/delete",
+    "id": "${articleId}"
+  }`
+
+  const newArticleDeletePostEvent = createNewPostEvent(otherUser, articleDeleteContent, PUBLICATION_TAG)
+  handleNewPost(newArticleDeletePostEvent)
+
+  // check that the article is not deleted
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
+
+  clearStore()
+})
+
+test("An account can update an article in a publication where the account has `article/update` permissions", () => {
+  const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
+  const publicationTitle = "My First Publication"
+  const publicationContent = `{
+    "action": "publication/create",
+    "title": "${publicationTitle}"
+  }`
+
+  const newPublicationPostEvent = createNewPostEvent(user, publicationContent, PUBLICATION_TAG)
+  const publicationId = getPublicationId(newPublicationPostEvent)
+  handleNewPost(newPublicationPostEvent)
+
+  assert.fieldEquals(PUBLICATION_ENTITY_TYPE, publicationId, "id", publicationId)
+
+  const articleTitle = "My First Blog Post"
+  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
+  const articleContent = `{
+    "action": "article/create",
+    "publicationId": "${publicationId}",
+    "article": "${article}",
+    "title": "${articleTitle}"
+  }`
+
+  const newArticlePostEvent = createNewPostEvent(user, articleContent, PUBLICATION_TAG)
+  const articleId = getArticleId(newArticlePostEvent)
+  handleNewPost(newArticlePostEvent)
+
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
   assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
 
   const updatedTitle = "My First Edited Blog Post"
@@ -137,22 +205,34 @@ test("An account can update a standalone article that it has created", () => {
   clearStore()
 })
 
-test("An account can NOT update a standalone article that it has NOT created", () => {
+test("An account can NOT update an article in a publication where the account has does not have any permissions", () => {
   const user = Address.fromString("0x89205A3A3b2A69De6Dbf7f01ED13B2108B2c43e7")
-  const title = "My First Blog Post"
-  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
-  const createArticleContent = `{
-    "action": "article/create",
-    "article": "${article}",
-    "title": "${title}"
+  const publicationTitle = "My First Publication"
+  const publicationContent = `{
+    "action": "publication/create",
+    "title": "${publicationTitle}"
   }`
 
-  const createArticlePostEvent = createNewPostEvent(user, createArticleContent, PUBLICATION_TAG)
-  const articleId = getArticleId(createArticlePostEvent)
-  handleNewPost(createArticlePostEvent)
+  const newPublicationPostEvent = createNewPostEvent(user, publicationContent, PUBLICATION_TAG)
+  const publicationId = getPublicationId(newPublicationPostEvent)
+  handleNewPost(newPublicationPostEvent)
 
-  // check that the article is in the store
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
+  assert.fieldEquals(PUBLICATION_ENTITY_TYPE, publicationId, "id", publicationId)
+
+  const articleTitle = "My First Blog Post"
+  const article = "QmbtLeBCvT1FW1Kr1JdFCPAgsVsgowg3zMJQS8eFrwPP2j"
+  const articleContent = `{
+    "action": "article/create",
+    "publicationId": "${publicationId}",
+    "article": "${article}",
+    "title": "${articleTitle}"
+  }`
+
+  const newArticlePostEvent = createNewPostEvent(user, articleContent, PUBLICATION_TAG)
+  const articleId = getArticleId(newArticlePostEvent)
+  handleNewPost(newArticlePostEvent)
+
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
   assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
 
   const updatedTitle = "My First Edited Blog Post"
@@ -169,8 +249,8 @@ test("An account can NOT update a standalone article that it has NOT created", (
   const newUpdatePostEvent = createNewPostEvent(otherUser, updateArticleContent, PUBLICATION_TAG)
   handleNewPost(newUpdatePostEvent)
 
-  // check that the article is NOT updated
-  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", title)
+  // check that the article is updated
+  assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "title", articleTitle)
   assert.fieldEquals(ARTICLE_ENTITY_TYPE, articleId, "article", article)
 
   clearStore()
